+++ src/main/java/com/gotreaux/aoc/puzzles/year2022/day2/Round.java
@@ -1,17 +1,17 @@
 package com.gotreaux.aoc.puzzles.year2022.day2;
 
 public record Round(Hand opponentHand, Hand strategyHand) {
-    public long getScore() {
+    public int getScore() {
         if (opponentHand == strategyHand) {
-            return 3L;
+            return 3;
         }
 
         if ((opponentHand == Hand.ROCK && strategyHand == Hand.PAPER)
                 || (opponentHand == Hand.PAPER && strategyHand == Hand.SCISSORS)
                 || (opponentHand == Hand.SCISSORS && strategyHand == Hand.ROCK)) {
-            return 6L;
+            return 6;
         }
 
-        return 0L;
+        return 0;
     }
 }
